Update Kernel Memory Allocation. #20220820

+++ Kernel/include/graphic/framebuffer.h
@@ -1,5 +1,5 @@
 #include <macros.h>
 
-void set_framebuffer(uint32_t address);
+void set_framebuffer(uint64_t address);
 
 void draw_rect();
